requirements.txt
# ── Core data / analysis (unchanged) ──────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
python-dateutil>=2.8.2

# ── OpenBB market data ────────────────────────────────────────────────────
# Install the base package + the yfinance provider extension.
# For other providers (FMP, Polygon, etc.) install the matching extension:
#   pip install openbb-fmp openbb-polygon openbb-intrinio
openbb>=4.2.0
openbb-yfinance>=0.2.0

# ── Interactive Brokers paper trading ─────────────────────────────────────
# Requires TWS or IB Gateway running locally with API connections enabled.
ib_insync>=0.9.86

# ── World Monitor HTTP client ─────────────────────────────────────────────
# httpx is used for the World Monitor live API; not needed in stub mode.
httpx>=0.27.0
